add structure validator
