+++ .circleci/config.yml
@@ -34,7 +34,7 @@ jobs:
           name: Validate Files and Yaml
           command: |
               ./Tests/scripts/pyflake.sh
-              python ./Tests/scripts/validate_files_structure.py -c true
+              python ./Tests/scripts/validate_files.py -c true
       - run:
           name: Configure Test Filter
           command: |
