+++ .hooks/pre-commit
@@ -1,9 +1,11 @@
 #!/bin/bash
 
+# Running python linter
+./Tests/scripts/pyflake.sh
+
 # validating that each modified file has a valid schema, release notes, proper prefix & suffix
 echo "Validating files"
-python Tests/scripts/validate_files_structure.py
-flake8
+python Tests/scripts/validate_files.py
 
 if [[ $? -ne 0 ]]
   then
