+++ Tests/scripts/hook_validations/tests/structure_test.py
@@ -1,0 +1,5 @@
+from Tests.scripts.hook_validations.structure import StructureValidator
+
+
+def test():
+    pass
